requirements.txt
fastapi
uvicorn
rasterio
groq
requests
python-dotenv